Move breadcrumb styles to CSS module + move tests their own file
