+++ src/components/Breadcrumb.tsx
@@ -1,6 +1,8 @@
 import type { ReactNode } from 'react'
 import { useConfig } from '../hooks/useConfig.js'
 import type { Source } from '../lib/sources/types.js'
+import { cn } from '../lib/utils.js'
+import styles from '../styles/Breadcrumb.module.css'
 
 interface BreadcrumbProps {
   source: Source,
@@ -11,10 +13,10 @@ interface BreadcrumbProps {
  * Breadcrumb navigation
  */
 export default function Breadcrumb({ source, children }: BreadcrumbProps) {
-  const { routes } = useConfig()
+  const { routes, customClass } = useConfig()
 
-  return <nav className='top-header top-header-divided'>
-    <div className='path'>
+  return <nav className={cn(styles.breadcrumb, customClass?.breadcrumb)}>
+    <div className={cn(styles.path, customClass?.path)}>
       {source.sourceParts.map((part, depth) =>
         <a href={routes?.getSourceRouteUrl?.({ sourceId: part.sourceId }) ?? ''} key={depth}>{part.text}</a>
       )}
