+++ src/styles/Breadcrumb.module.css
@@ -1,0 +1,48 @@
+.breadcrumb {
+  /* top navbar */
+  align-items: center;
+  display: flex;
+  font-size: 18px;
+  height: 32px;
+  justify-content: space-between;
+  min-height: 32px;
+  padding-left: 20px;
+  padding-right: 10px;
+  border-bottom: 1px solid #ddd;
+  background: #eee;
+  /* TODO(SL): forbid overflow? */
+
+  h1 {
+    font-size: 18px;
+    margin: 4px 0 0 0; /* top */
+    user-select: none;
+  }
+}
+
+/* file path */
+.path {
+  margin: 0 2px;
+  margin-right: 4px;
+  min-width: 0;
+  overflow: auto;
+  /* TODO(SL): forbid wrap + use an ellipsis instead? */
+
+  &::-webkit-scrollbar {
+    display: none;
+  }
+  a {
+    color: #222622;
+    font-family: "Courier New", Courier, monospace;
+    font-weight: 600;
+    font-size: 18px;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+    text-decoration-thickness: 1px;
+  }
+  /* hide all but the last path link on small screens */
+  @media (max-width: 360px) {
+    & a:not(:last-child) {
+      display: none;
+    }
+  }
+}
